Create test doc.md
Test to see how it affects my environment. 

Curious to see how this affects the .md file extension. 

What do I think will happen? It will add another file to the git "my-first-project" assuming I don't delete when prompted. 

How many forks will be associated with this repo? It depends on if I commit directly to the main branch or create a pull request. 

I will create a pull request and not delete the branch.